♻️ Move all amd64 HAL stuff into HAL trait
